functionality related to querying cross-chain messages info in a separate module
