+++ apps/indexer/lib/indexer/fetcher/arbitrum/rollup_messages_catchup.ex
@@ -61,7 +61,7 @@ defmodule Indexer.Fetcher.Arbitrum.RollupMessagesCatchup do
 
   import Indexer.Fetcher.Arbitrum.Utils.Logging, only: [log_warning: 1]
 
-  alias Indexer.Fetcher.Arbitrum.Utils.Db
+  alias Indexer.Fetcher.Arbitrum.Utils.Db.Common, as: DbCommon
   alias Indexer.Fetcher.Arbitrum.Workers.HistoricalMessagesOnL2
 
   require Logger
@@ -153,7 +153,7 @@ defmodule Indexer.Fetcher.Arbitrum.RollupMessagesCatchup do
       end
 
     new_data =
-      case Db.closest_block_after_timestamp(time_of_start) do
+      case DbCommon.closest_block_after_timestamp(time_of_start) do
         {:ok, block} ->
           Process.send(self(), :init_worker, [])
