+++ apps/indexer/lib/indexer/fetcher/arbitrum/tracking_batches_statuses.ex
@@ -53,8 +53,10 @@ defmodule Indexer.Fetcher.Arbitrum.TrackingBatchesStatuses do
   import Indexer.Fetcher.Arbitrum.Utils.Helper, only: [increase_duration: 2]
 
   alias EthereumJSONRPC.Arbitrum, as: ArbitrumRpc
+  alias Indexer.Fetcher.Arbitrum.Utils.Db
+  alias Indexer.Fetcher.Arbitrum.Utils.Db.Messages, as: DbMessages
+  alias Indexer.Fetcher.Arbitrum.Utils.Rpc
   alias Indexer.Helper, as: IndexerHelper
-  alias Indexer.Fetcher.Arbitrum.Utils.{Db, Rpc}
 
   require Logger
 
@@ -174,8 +176,8 @@ defmodule Indexer.Fetcher.Arbitrum.TrackingBatchesStatuses do
 
     new_confirmations_start_block = Db.l1_block_of_latest_confirmed_block(l1_start_block)
 
-    new_executions_start_block = Db.l1_block_to_discover_latest_execution(l1_start_block)
-    historical_executions_end_block = Db.l1_block_to_discover_earliest_execution(l1_start_block - 1)
+    new_executions_start_block = DbMessages.l1_block_to_discover_latest_execution(l1_start_block)
+    historical_executions_end_block = DbMessages.l1_block_to_discover_earliest_execution(l1_start_block - 1)
 
     {lowest_batch, missing_batches_end_batch} = Db.get_min_max_batch_numbers()
